+++ src/main/java/racingcar/Application.java
@@ -1,12 +1,13 @@
 package racingcar;
 
+import racingcar.controller.RacingGameController;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class Application {
 
     public static void main(final String[] args) {
-        final RacingGame racingGame = new RacingGame(new InputView(), new OutputView());
-        racingGame.start();
+        final RacingGameController racingGameController = new RacingGameController(new InputView(), new OutputView());
+        racingGameController.start();
     }
 }
