+++ src/main/java/racingcar/domain/Cars.java
@@ -5,6 +5,9 @@ import java.util.List;
 
 public class Cars {
 
+    private static final String DELIMITER = ",";
+    private static final int INITIAL_CAR_POSITION = 0;
+
     private final List<Car> cars;
 
     private Cars(final List<Car> cars) {
@@ -12,7 +15,7 @@ public class Cars {
     }
 
     public static Cars from(final String carNames) {
-        final List<Car> cars = Arrays.stream(carNames.split(","))
+        final List<Car> cars = Arrays.stream(carNames.split(DELIMITER))
                 .map(Car::new)
                 .toList();
         return new Cars(cars);
@@ -29,7 +32,7 @@ public class Cars {
         final int maxPosition = cars.stream()
                 .mapToInt(Car::getPosition)
                 .max()
-                .orElse(0);
+                .orElse(INITIAL_CAR_POSITION);
         return cars.stream()
                 .filter(car -> car.getPosition() == maxPosition)
                 .map(Car::getName)
