+++ src/main/java/racingcar/domain/vo/Position.java
@@ -1,6 +1,7 @@
 package racingcar.domain.vo;
 
 public record Position(int value) {
+    
     public Position increase() {
         return new Position(value + 1);
     }
